Add files via upload
The purpose of this project is to build a deep learning model that can accurately distinguish between drones and birds. This distinction is crucial in various applications, such as surveillance systems, where differentiating between these two objects is necessary for accurate threat detection and monitoring.

Key Features:
Model Architecture: The CNN model was designed with multiple convolutional layers, pooling layers, and fully connected layers to effectively capture spatial hierarchies and features within the images.
Data Augmentation: Techniques such as rotation, flipping, and zooming were applied to the training dataset to improve the model's generalization capability and avoid overfitting.
Performance: The final model achieved an accuracy of 96.33% on the test dataset, demonstrating its robustness in distinguishing between drones and birds.